requirements.txt
# StockFlowML - Level 1 Dependencies
# Production-ready MLOps pipeline for stock trend prediction

# Data & Finance
yfinance==0.2.32
pandas==2.1.4
numpy==1.26.2

# Machine Learning
scikit-learn==1.3.2
xgboost==2.0.3
imbalanced-learn==0.11.0

# MLOps
dvc==3.37.0
mlflow==2.9.2

# Monitoring (Level 2 scaffold)
evidently==0.4.11

# Feature Store (Level 3 scaffold)
feast==0.35.0

# API (Level 3 scaffold)  
fastapi==0.105.0
uvicorn[standard]==0.25.0
pydantic==2.5.3

# Utilities
PyYAML==6.0.1
python-dateutil==2.8.2
tqdm==4.66.1
python-dotenv==1.0.0
requests==2.31.0

# Visualization (EDA only)
matplotlib==3.8.2
seaborn==0.13.0
